refactor(engine): relocate 2 pure collector helpers to rts-engine
Pré-requisito do cluster (function/collections/proxy/error/string/reflect): dois pub fns com tipo Rust que o cluster consome (não dá extern-decl) saem do collector do rts-runtime pro motor.

- format_js_number(f64)->String -> rts_engine::numfmt (puro). Re-export em collector/string_pool pros call-sites (json, collections/vec) seguirem.
- stack_for_handle(u64)->Option<String> + thread-local ERR_STACKS -> rts_engine::collector::err_stack. ERROR_CLEAR chama err_stack::record; o slot de erro pendente (ERROR_SLOT) fica no runtime. Re-export em collector/error pro globals/error seguir.

Gate: cargo build --release limpo; rts.exe test 1710/1710; AOT smoke (error.stack + Number.toString via engine, JIT==AOT) OK.
